feat: Reestructure project.
## Main Changes in Structure

1. **Controller Reorganization (`src/controller/`)**:
   - **Removal of `opensearch.rs`**: The monolithic OpenSearch file was split into smaller modules for better separation of concerns.
   - **New modules created**:
     - `api_rest.rs`: Contains all REST API routes and handlers (previously inline in `server.rs`).
     - `kafka_consumer.rs`: Handles Kafka consumer configuration and creation.
     - `kafka_queue.rs`: Manages message consumption from the Kafka queue.
     - `opensearch_client.rs`: Encapsulates OpenSearch client creation.
     - `opensearch_document.rs`: Functions related to document operations (create, delete, search).
     - `opensearch_index.rs`: Functions for index management (create, delete, check existence).
   - **Update to `mod.rs`**: Reflects the new modules and removes obsolete references.

2. **Main Server Simplification (`src/bin/server.rs`)**:
   - Removed the inline `routes()` and `consume_queue()` functions, moving them to dedicated modules.
   - The `main()` function is now cleaner and focused on initializing the server and asynchronous tasks.
   - Updated imports to use the new modules (`api_rest::routes`, `kafka_queue::consume_queue`).

3. **Environment Configuration Management**:
   - **Removal of `docker/.env`**: Docker-specific configuration file removed to avoid duplication.
   - **Creation of `env_template.env`**: New template file with default values (e.g., `ALLOWED_ORIGIN=*`, `KAFKA_ENABLED=Y`).
   - **Update to `.gitignore`**: Added `.env` and `/docker/.env` to ignore sensitive configuration files.
   - **Update to `README.md`**: Added instructions to copy `env_template.env` to `.env` and edit it before running the project.

4. **Updates in Models (`src/models/`)**:
   - Import changes in several files (`browser.rs`, `form.rs`, `generic.rs`, `menu.rs`, `menu_item.rs`, `menu_tree.rs`, `mod.rs`, `process.rs`, `role.rs`, `window.rs`) to use the new `controller` modules (e.g., `opensearch_document` instead of `opensearch`).

## Benefits of the Restructuring

- **Better Maintainability**: Code is more modular, making it easier to edit and debug specific components (e.g., REST API vs. Kafka vs. OpenSearch).
- **Separation of Responsibilities**: Each module has a clear purpose (OpenSearch client, Kafka consumer, API routes).
- **Safer Configuration**: Use of templates for environments, avoiding commits of sensitive configurations.
- **Readability**: The `server.rs` file is more concise and easier to understand.
- **Scalability**: Facilitates adding new features without overloading existing files.

## Affected Files

- **New**: `env_template.env`, `src/controller/api_rest.rs`, `src/controller/kafka_consumer.rs`, `src/controller/kafka_queue.rs`, `src/controller/opensearch_client.rs`, `src/controller/opensearch_document.rs`, `src/controller/opensearch_index.rs`.
- **Modified**: `.gitignore`, `README.md`, `src/bin/server.rs`, `src/controller/mod.rs`, and several in `src/models/`.
- **Deleted**: `docker/.env`, `src/controller/opensearch.rs`.
