+++ .gitignore
@@ -8,3 +8,7 @@ Cargo.lock
 
 # These are backup files generated by rustfmt
 **/*.rs.bk
+
+# Environment file
+.env
+/docker/.env
