--- docker/.env
@@ -1,9 +1,0 @@
-RUST_LOG=info
-PORT=7878
-ALLOWED_ORIGIN="allowed_origin"
-KAFKA_ENABLED="kafka_enabled"
-KAFKA_QUEUES="kafka_queues"
-KAFKA_HOST="kafka_host"
-KAFKA_GROUP="kafka_group"
-OPENSEARCH_URL="opensearch_url"
-VERSION="1.0.0-dev"
