+++ src/controller/mod.rs
@@ -1,2 +1,6 @@
-pub mod kafka;
-pub mod opensearch;
+pub mod api_rest;
+pub mod kafka_consumer;
+pub mod kafka_queue;
+pub mod opensearch_client;
+pub mod opensearch_document;
+pub mod opensearch_index;
