+++ src/models/window.rs
@@ -3,7 +3,8 @@ use salvo::prelude::*;
 use serde_json::{json, Value};
 use std::{io::ErrorKind, io::Error};
 
-use crate::{controller::opensearch::{IndexDocument, find, get_by_id}, models::{generic::DependendField, get_index_name}};
+use crate::{controller::opensearch_document::{IndexDocument, find, get_by_id}, models::{generic::DependendField, get_index_name}};
+
 
 #[derive(Deserialize, Extractible, Debug, Clone)]
 #[salvo(extract(default_source(from = "body")))]
